+++ authorizations/apps.py
@@ -5,4 +5,5 @@ class AuthConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'authorizations'
 
-
+    def ready(self):
+        import authorizations.security.signals